LayoutTests/imported/w3c/web-platform-tests/css/css-contain/contain-size-multicol-as-flex-item-expected.txt

There should be a green square below.


FAIL #flex 1 assert_equals:
<div id="flex">
  <div id="multicol" data-expected-width="100" data-expected-height="100">
    <div style="width:1000px; height:1px;"></div>
  </div>
</div>
width expected 100 but got 784
